gRPC/pyproject.toml
[tool.poetry]
name = "grpc"
version = "0.1.0"
description = "gRPC example project with python"
authors = ["Katzper <kacper.wozni@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.58.0"
grpcio-tools = "^1.58.0"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
